backend/requirements.txt
fastapi>=0.121.3
uvicorn>=0.38.0
mysql-connector-python>=9.5.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.10.1
pydantic-settings>=2.12.0
email-validator>=2.3.0
requests>=2.32.5








